# [c++,python] Correct SOMA type for collection members