--- a/apis/r/R/SOMACollectionBase.R
+++ b/apis/r/R/SOMACollectionBase.R
@@ -211,6 +211,16 @@ SOMACollectionBase <- R6::R6Class(
         SOMAArray = TileDBArray$new,
         GROUP = TileDBGroup$new,
         SOMAGroup = TileDBGroup$new,
+        SOMADataFrame = TileDBArray$new,
+        SOMADenseNDArray = TileDBArray$new,
+        SOMAGeometryDataFrame = TileDBArray$new,
+        SOMAPointCloudDataFrame = TileDBArray$new,
+        SOMASparseNDArray = TileDBArray$new,
+        SOMACollection = TileDBGroup$new,
+        SOMAExperiment = TileDBGroup$new,
+        SOMAMeasurement = TileDBGroup$new,
+        SOMAScene = TileDBGroup$new,
+        SOMAMultiscaleImage = TileDBGroup$new,
         stop(sprintf("Unknown member TileDB type: %s", type), call. = FALSE)
       )