--- a/libtiledbsoma/src/soma/soma_collection.cc
+++ b/libtiledbsoma/src/soma/soma_collection.cc
@@ -90,7 +90,8 @@ std::shared_ptr<SOMACollection> SOMACollection::add_new_collection(
     // in addition to returning the SOMA object to the user.
     std::shared_ptr<SOMACollection> member = SOMACollection::open(
         uri, OpenMode::read, ctx, timestamp);
-    this->set(std::string(uri), uri_type, std::string(key), "SOMAGroup");
+    this->set(
+        std::string(uri), uri_type, std::string(key), member->type().value());
     children_[std::string(key)] = member;
     return member;
 }
@@ -116,7 +117,8 @@ std::shared_ptr<SOMAExperiment> SOMACollection::add_new_experiment(
     // in addition to returning the SOMA object to the user.
     std::shared_ptr<SOMAExperiment> member = SOMAExperiment::open(
         uri, OpenMode::read, ctx, timestamp);
-    this->set(std::string(uri), uri_type, std::string(key), "SOMAGroup");
+    this->set(
+        std::string(uri), uri_type, std::string(key), member->type().value());
     children_[std::string(key)] = member;
     return member;
 }
@@ -142,7 +144,8 @@ std::shared_ptr<SOMAMeasurement> SOMACollection::add_new_measurement(
     // in addition to returning the SOMA object to the user.
     std::shared_ptr<SOMAMeasurement> member = SOMAMeasurement::open(
         uri, OpenMode::read, ctx, timestamp);
-    this->set(std::string(uri), uri_type, std::string(key), "SOMAGroup");
+    this->set(
+        std::string(uri), uri_type, std::string(key), member->type().value());
     children_[std::string(key)] = member;
     return member;
 }
@@ -168,7 +171,8 @@ std::shared_ptr<SOMADataFrame> SOMACollection::add_new_dataframe(
     // in addition to returning the SOMA object to the user.
     std::shared_ptr<SOMADataFrame> member = SOMADataFrame::open(
         uri, OpenMode::read, ctx, timestamp);
-    this->set(std::string(uri), uri_type, std::string(key), "SOMAArray");
+    this->set(
+        std::string(uri), uri_type, std::string(key), member->type().value());
     children_[std::string(key)] = member;
     return member;
 }
@@ -194,7 +198,8 @@ std::shared_ptr<SOMADenseNDArray> SOMACollection::add_new_dense_ndarray(
     // in addition to returning the SOMA object to the user.
     std::shared_ptr<SOMADenseNDArray> member = SOMADenseNDArray::open(
         uri, OpenMode::read, ctx, timestamp);
-    this->set(std::string(uri), uri_type, std::string(key), "SOMAArray");
+    this->set(
+        std::string(uri), uri_type, std::string(key), member->type().value());
     children_[std::string(key)] = member;
     return member;
 }
@@ -220,7 +225,8 @@ std::shared_ptr<SOMASparseNDArray> SOMACollection::add_new_sparse_ndarray(
     // in addition to returning the SOMA object to the user.
     std::shared_ptr<SOMASparseNDArray> member = SOMASparseNDArray::open(
         uri, OpenMode::read, ctx, timestamp);
-    this->set(std::string(uri), uri_type, std::string(key), "SOMAArray");
+    this->set(
+        std::string(uri), uri_type, std::string(key), member->type().value());
     children_[std::string(key)] = member;
     return member;
 }