**Issue and/or context:**

[[sc-65770](https://app.shortcut.com/tiledb-inc/story/65770/python-somacollection-members-returns-incorrect-soma-type-for-collection-members)]

**Changes:**

Previously, the code would return `SOMAArray` for `tiledb::Object::Array` and `SOMAGroup` for `tiledb::Object::Group`. Now we use `SOMAArray::open` and `SOMAGroup::open` and use `SOMAObject::type` to get the derived class.