@bkmartinjr 2025-04-22
> I want to highlight that this change will make opening groups slower because we now open every member in the group in order to read the SOMA type metadata.

Do we have a ballpark on how much slower?