@nguyenv 2025-04-23
This PR introduces a VERY significant slowdown.

I created this group with 10000 members:
```
    df = pd.DataFrame(data={"soma_joinid": [0, 1], "col1": [1, 2], "col2": [3, 4]})
    schema = pa.Schema.from_pandas(df)    
    uri = "lots-of-members.tdb"

    with soma.Collection.create(uri) as col:
        for i in range(10000):
            col.add_new_dataframe(f"df{i}", schema=schema)
```

The benchmarking code is simply:
```
start = time.time()
soma.Collection.open("lots-of-members.tdb")
end = time.time()
print(f"Collection.open took {end - start:.6f} seconds")
```

Without these changes:
```
Successfully installed tiledbsoma-1.16.1
(py311) vivian@mangonada:~/tiledb-bugs$ python lots-of-members.py
Collection.open took 0.147007 seconds
(py311) vivian@mangonada:~/tiledb-bugs$ python lots-of-members.py
Collection.open took 0.133340 seconds
(py311) vivian@mangonada:~/tiledb-bugs$ python lots-of-members.py
Collection.open took 0.109550 seconds
(py311) vivian@mangonada:~/tiledb-bugs$ python lots-of-members.py
Collection.open took 0.121464 seconds
(py311) vivian@mangonada:~/tiledb-bugs$ python lots-of-members.py
Collection.open took 0.103636 seconds
```

With these changes:
```
(py311) vivian@mangonada:~/tiledb-bugs$ python lots-of-members.py
Collection.open took 15.714127 seconds
(py311) vivian@mangonada:~/tiledb-bugs$ python lots-of-members.py
Collection.open took 13.562405 seconds
(py311) vivian@mangonada:~/tiledb-bugs$ python lots-of-members.py
Collection.open took 14.042180 seconds
(py311) vivian@mangonada:~/tiledb-bugs$ python lots-of-members.py
Collection.open took 13.675979 seconds
(py311) vivian@mangonada:~/tiledb-bugs$ python lots-of-members.py
Collection.open took 13.978573 seconds
```

I am running this locally, so this would be even worse on cloud.